+++ resources/views/file/create.blade.php
@@ -12,7 +12,7 @@
                 value="{{ $context->valueOf('title') }}" width2="4" />
                 <x-laraknife.forms.combobox position="first" name="visibility_scope" label="Visibility" :options="$optionsVisibility"
                 width2="4" />
-            <x-laraknife.forms.file position="first" name="file" label="File" width2="4" rows="4" />
+            <x-laraknife.forms.file position="first" name="file" label="File" width2="4" />
             <x-laraknife.forms.string position="last" name="filename" label="Filename"
                 value="{{ $context->valueOf('filename') }}" width2="4" placeholder="Overwrites the upload name (if not empty)"/>
             <x-laraknife.forms.text position="alone" name="description" label="Description"